requirements.txt
Flask==2.3.3
numpy==1.26.4
Pillow==10.4.0
h5py==3.11.0
efficientnet==1.1.1
tensorflow-cpu==2.15.0
gunicorn==21.2.0